AutoArmor Recode，合并NoJump && NoClick Delay

+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/other/Spammer.kt
@@ -62,7 +62,7 @@ object Spammer : Module() {
             return
         }
 
-        if (msTimer.hasTimePassed(delay)) {
+        if (msTimer.hasTimePassed(delay.toLong())) {
             mc.thePlayer.sendChatMessage(when (modeValue.get().lowercase()) {
                 "insult" -> {
                     replaceAbuse(Insult.getRandomOne())
